+++ CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Add `StoreInterface::remove()` method
+ * [BC BREAK] Make `Indexer` stateless - removed `$source` constructor parameter and `withSource()` method
+ * Add `ConfiguredIndexer` decorator for pre-configuring default sources
 
 0.3
 ---